Re-organized the CPU into it's own file :godmode:
